fix bug in main time loop
Changed the main time loop to be controlled by an integer index, rather than by a condition on real numbers. The latter was creating an issue found by Liuwei Xu (UCLA) when running POST/matlab/sample_fault.m on EXAMPLES/2.5D_inplane: in the fault outputs, the number of time steps in the header file and in the data file were inconsistent (the former was = the latter + 1).